civicswipe/backend/requirements.txt
# CivicSwipe Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
cryptography==42.0.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
beautifulsoup4==4.12.3
lxml==5.1.0

# Task Queue
celery==5.3.4
redis==5.0.1

# AI/ML
openai==1.7.2
anthropic==0.77.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.39.1

# Geocoding
geopy==2.4.1

# PDF Processing
PyPDF2==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0

# Development
black==23.12.1
ruff==0.1.9
mypy==1.8.0
